// FoodTrackerClient/food_tracker/src/ClientGenerator/generated/models/Meal.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Meal Tracker API
 * API documentation for the Meal Tracker Service
 *
 * The version of the OpenAPI document: v1
 * Contact: your.email@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FoodItem } from './FoodItem';
import {
    FoodItemFromJSON,
    FoodItemFromJSONTyped,
    FoodItemToJSON,
    FoodItemToJSONTyped,
} from './FoodItem';

/**
 * 
 * @export
 * @interface Meal
 */
export interface Meal {
    /**
     * 
     * @type {string}
     * @memberof Meal
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Meal
     */
    name?: string | null;
    /**
     * 
     * @type {Array<FoodItem>}
     * @memberof Meal
     */
    foodItems?: Array<FoodItem> | null;
    /**
     * 
     * @type {number}
     * @memberof Meal
     */
    calories?: number;
}

/**
 * Check if a given object implements the Meal interface.
 */
export function instanceOfMeal(value: object): value is Meal {
    return true;
}

export function MealFromJSON(json: any): Meal {
    return MealFromJSONTyped(json, false);
}

export function MealFromJSONTyped(json: any, ignoreDiscriminator: boolean): Meal {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'foodItems': json['foodItems'] == null ? undefined : ((json['foodItems'] as Array<any>).map(FoodItemFromJSON)),
        'calories': json['calories'] == null ? undefined : json['calories'],
    };
}

export function MealToJSON(json: any): Meal {
    return MealToJSONTyped(json, false);
}

export function MealToJSONTyped(value?: Meal | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'foodItems': value['foodItems'] == null ? undefined : ((value['foodItems'] as Array<any>).map(FoodItemToJSON)),
        'calories': value['calories'],
    };
}